way of modelling persons and roles changed
Now use the EntityInRole approach. Have also separated out production and non-production organisations, to allow a clear mapping to the Productie and Namen schemes in the GTAA. On inconsistency still exists: Executive (uitvoerende) is modelled as using an Entity, which would imply it can be a person, a production organisation or a non-production organisation. In fact, it can only be a person or a non-production organisation, but to model this we would have to introduce some concept that unifies a person and a non-production organisation, which is quite artificial and would make the schema harder to understand. Have also changed hasProductionCountry to us a Location instead of a ProductionLand, to harmonise how locations are indicated.