Fix B008 (function calls in argument defaults)
B008 Do not perform function calls in argument defaults.  The call
is performed only once at function definition time. All calls to your
function will reuse the result of that definition-time function call.
If this is intended, assign the function call to a module-level variable
and use that variable as a default value.